% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorhull.r
\name{get_theme_colors_from_image}
\alias{get_theme_colors_from_image}
\title{Create a theme for an image}
\usage{
get_theme_colors_from_image(path)
}
\arguments{
\item{path}{Image path}
}
\value{
Image color theme
}
\description{
Create a theme for an image
}
